# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BTCFi API",
    "description": "Agent-native Bitcoin data, intelligence, BTCFi & security API via x402 micropayments. No tokens. No subscriptions. Payment IS authentication.",
    "version": "2.0.0",
    "contact": {
      "name": "AI Indigo",
      "url": "https://aiindigo.com",
      "email": "contact@aiindigo.com"
    },
    "license": { "name": "MIT" }
  },
  "servers": [
    { "url": "https://btcfi.aiindigo.com", "description": "Production" },
    { "url": "http://localhost:3001", "description": "Development" }
  ],
  "paths": {
...
    "/api/v1/address/{addr}/txs": {
      "get": {
        "summary": "Address Transactions",
        "tags": ["Address"],
        "parameters": [{ "name": "addr", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Transaction history" },
          "402": { "$ref": "#/components/responses/PaymentRequired" }
        }
      }
    },
...
    "/api/v1/block/latest": {
      "get": {
        "summary": "Latest Blocks",
        "tags": ["Block"],
        "parameters": [{ "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } }],
        "responses": {
          "200": { "description": "Recent blocks" },
          "402": { "$ref": "#/components/responses/PaymentRequired" }
        }
      }
    },
...
    "/api/v1/staking/status": {
      "get": {
        "summary": "Staking Status",
        "tags": ["Staking"],
        "parameters": [{ "name": "address", "in": "query", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Staking tier info" } }
      }
    },
...
  },
  "components": {
    "responses": {
      "PaymentRequired": {
        "description": "Payment required. Returns x402 payment instructions for Base (USDC) and Solana (USDC via NLx402).",
        "headers": {
          "X-Payment-Required": { "schema": { "type": "string" } },
          "X-Payment-Amount": { "schema": { "type": "string" } },
          "X-Payment-Networks": { "schema": { "type": "string" } }
        }
      }
    },
    "securitySchemes": {
      "x402Payment": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Payment",
        "description": "x402 payment proof (Base64 encoded)"
      },
      "walletSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Signature",
        "description": "Wallet signature for higher rate limits"
      }
    }
  },
...
}